Do not run NDN checks on A8C proxied requests
We should trust and eventually not report on proxied requests coming to the site.

This commit is hooking the `ndn_run_for_current_user` filter and returns `false` (for skipping the check), in case the `A8C_PROXIED_REQUEST` constant is defined and set to `true`. The new filter returns default value (currently determined as `$is_privileged_user = ! is_proxied_automattician() && current_user_can( 'edit_posts' );`) in case the request is not proxied.